+++ packages/typespec-python/src/emitter.ts
@@ -64,12 +64,6 @@ interface HttpServerParameter {
     param: ModelProperty;
 }
 
-const defaultOptions = {
-    "package-version": "1.0.0b1",
-    "generate-packaging-files": true,
-    "unbranded": false,
-};
-
 export function getModelsMode(context: SdkContext): "msrest" | "dpg" | "none" {
     const specifiedModelsMode = context.emitContext.options["models-mode"];
     if (specifiedModelsMode) {
@@ -83,44 +77,48 @@ export function getModelsMode(context: SdkContext): "msrest" | "dpg" | "none" {
     return "dpg";
 }
 
-function addDefaultCalculatedOptions(
-    sdkContext: SdkContext,
-    options: PythonEmitterOptions & InternalPythonEmitterOptions,
-    yamlMap: Record<string, any>,
-) {
+function addDefaultOptions(sdkContext: SdkContext) {
+    const defaultOptions = {
+        "package-version": "1.0.0b1",
+        "generate-packaging-files": true,
+        "flavor": undefined,
+    };
+    sdkContext.emitContext.options = {
+        ...defaultOptions,
+        ...sdkContext.emitContext.options,
+    };
+    const options = sdkContext.emitContext.options;
     options["models-mode"] = getModelsMode(sdkContext);
     if (options["generate-packaging-files"]) {
         options["package-mode"] = sdkContext.arm ? "azure-mgmt" : "azure-dataplane";
     }
     if (!options["package-name"]) {
-        options["package-name"] = yamlMap["namespace"].replace(/\./g, "-");
+        options["package-name"] = getClientNamespaceStringHelper(sdkContext).replace(/\./g, "-");
+    }
+    if (options.flavor !== "azure") {
+        // if they pass in a flavor other than azure, we want to ignore the value
+        options.flavor = undefined;
+    }
+    if (!options.flavor && sdkContext.emitContext.emitterOutputDir.includes("azure/")) {
+        options.flavor = "azure";
     }
 }
-
-interface InternalPythonEmitterOptions {
-    "package-mode"?: string;
-}
-
 export async function $onEmit(context: EmitContext<PythonEmitterOptions>) {
     const program = context.program;
-    const resolvedOptions: PythonEmitterOptions & InternalPythonEmitterOptions = {
-        ...defaultOptions,
-        ...context.options,
-    };
-
     const sdkContext = createSdkContext(context, "@azure-tools/typespec-python");
     const clients = listClients(sdkContext);
+    addDefaultOptions(sdkContext);
     const root = await resolveModuleRoot(program, "@autorest/python", dirname(fileURLToPath(import.meta.url)));
     const outputDir = context.emitterOutputDir;
     const yamlMap = emitCodeModel(sdkContext, clients);
     const yamlPath = await saveCodeModelAsYaml("typespec-python-yaml-map", yamlMap);
-    addDefaultCalculatedOptions(sdkContext, resolvedOptions, yamlMap);
     const commandArgs = [
         `${root}/run-python3.js`,
         `${root}/run_cadl.py`,
         `--output-folder=${outputDir}`,
         `--cadl-file=${yamlPath}`,
     ];
+    const resolvedOptions = sdkContext.emitContext.options;
     if (resolvedOptions["packaging-files-config"]) {
         const keyValuePairs = Object.entries(resolvedOptions["packaging-files-config"]).map(([key, value]) => {
             return `${key}:${value}`;
@@ -866,7 +864,7 @@ function emitCredentialParam(context: SdkContext, namespace: Namespace): Record<
                     types: credential_types,
                 };
             }
-            const service = context.emitContext.options["unbranded"] ? "cloud service" : "Azure";
+            const service = context.emitContext.options.flavor ? "Azure" : "cloud service";
             return {
                 type: getType(context, type),
                 optional: false,
@@ -952,7 +950,7 @@ function getNamespace(context: SdkContext, clientName: string): string {
     // We get client namespaces from the client name. If there's a dot, we add that to the namespace
     const submodule = clientName.split(".").slice(0, -1).join(".").toLowerCase();
     if (!submodule) {
-        return removeUnderscoresFromNamespace(getClientNamespaceString(context)!.toLowerCase());
+        return getClientNamespaceStringHelper(context);
     }
     return removeUnderscoresFromNamespace(submodule);
 }
@@ -965,8 +963,12 @@ function getNamespaces(context: SdkContext): Set<string> {
     return namespaces;
 }
 
+function getClientNamespaceStringHelper(context: SdkContext): string {
+    return removeUnderscoresFromNamespace(getClientNamespaceString(context)?.toLowerCase());
+}
+
 function emitCodeModel(sdkContext: SdkContext, clients: SdkClient[]) {
-    const clientNamespaceString = removeUnderscoresFromNamespace(getClientNamespaceString(sdkContext)?.toLowerCase());
+    const clientNamespaceString = getClientNamespaceStringHelper(sdkContext);
     // Get types
     const codeModel: Record<string, any> = {
         namespace: clientNamespaceString,