+++ packages/typespec-python/src/lib.ts
@@ -10,10 +10,10 @@ export interface PythonEmitterOptions {
     "package-pprint-name"?: string;
     "head-as-boolean"?: boolean;
     "models-mode"?: string;
-    "unbranded"?: boolean;
     "tracing"?: boolean;
     "company-name"?: string;
     "debug"?: boolean;
+    "flavor"?: "azure";
 }
 
 const EmitterOptionsSchema: JSONSchemaType<PythonEmitterOptions> = {
@@ -29,10 +29,10 @@ const EmitterOptionsSchema: JSONSchemaType<PythonEmitterOptions> = {
         "package-pprint-name": { type: "string", nullable: true },
         "head-as-boolean": { type: "boolean", nullable: true },
         "models-mode": { type: "string", nullable: true },
-        "unbranded": { type: "boolean", nullable: true },
         "tracing": { type: "boolean", nullable: true },
         "company-name": { type: "string", nullable: true },
         "debug": { type: "boolean", nullable: true },
+        "flavor": { type: "string", nullable: true },
     },
     required: [],
 };