+++ packages/autorest.python/autorest/codegen/serializers/builder_serializer.py
@@ -361,7 +361,10 @@ class _BuilderBaseSerializer(Generic[BuilderType]):  # pylint: disable=abstract-
                     "\n"
                 )
             )
-            docstring_type = param.docstring_type(async_mode=self.async_mode)
+            docstring_type = param.docstring_type(
+                async_mode=self.async_mode,
+                for_stream_response=getattr(builder, "for_stream_response", False),
+            )
             description_list.append(
                 f":{param.docstring_type_keyword} {param.client_name}: {docstring_type}"
             )
@@ -734,12 +737,8 @@ class _OperationSerializer(
         return kwargs
 
     def response_docstring(self, builder: OperationType) -> List[str]:
-        response_str = (
-            f":return: {builder.response_docstring_text(async_mode=self.async_mode)}"
-        )
-        rtype_str = (
-            f":rtype: {builder.response_docstring_type(async_mode=self.async_mode)}"
-        )
+        response_str = f":return: {builder.response_docstring_text(async_mode=self.async_mode, for_stream_response=builder.for_stream_response)}"  # pylint: disable=line-too-long
+        rtype_str = f":rtype: {builder.response_docstring_type(async_mode=self.async_mode, for_stream_response=builder.for_stream_response)}"  # pylint: disable=line-too-long
         return [
             response_str,
             rtype_str,
@@ -1112,12 +1111,12 @@ class _OperationSerializer(
         if response.headers:
             retval.append("")
         deserialize_code: List[str] = []
-        no_stream_logic = False
+        stream_logic = True
         if builder.has_stream_response:
             if isinstance(response.type, ByteArraySchema):
                 deserialized = f"{'await ' if self.async_mode else ''}response.read()"
             else:
-                no_stream_logic = True
+                stream_logic = False
                 if self.code_model.options["version_tolerant"]:
                     deserialized = "response.iter_bytes()"
                 else:
@@ -1171,7 +1170,7 @@ class _OperationSerializer(
                 deserialize_code.append("else:")
                 deserialize_code.append("    deserialized = None")
         if len(deserialize_code) > 0:
-            if builder.expose_stream_keyword and not no_stream_logic:
+            if builder.expose_stream_keyword and stream_logic:
                 retval.append("if _stream:")
                 retval.append("    deserialized = response.iter_bytes()")
                 retval.append("else:")